requirements.txt
numpy
certifi==2019.3.9
chardet==3.0.4
django-environ==0.4.5
idna==2.8
nltk==3.4
oauthlib==3.0.1
PySocks==1.6.8
python-twitter
requests==2.21.0
requests-oauthlib==1.2.0
Sastrawi==1.0.1
singledispatch==3.4.0.3
six==1.12.0
textblob==0.15.3
tweepy==3.7.0
urllib3==1.24.1
sklearn
